Move Utils file and implement secret call
