+++ src/main/java/com/google/sps/Utils.java
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package com.google.sps.servlets;
+package com.google.sps;
 
 import com.google.appengine.api.datastore.Entity;
 import com.google.gson.Gson;
@@ -25,7 +25,6 @@ import com.google.maps.errors.ApiException;
 import com.google.maps.model.DistanceMatrix;
 import com.google.maps.model.GeocodingResult;
 import com.google.maps.model.LatLng;
-import com.google.sps.Interactions;
 import java.io.IOException;
 import java.util.Comparator;
 import java.util.Map;
@@ -35,7 +34,17 @@ import javax.servlet.http.HttpServletRequest;
 public class Utils {
 
   private static final Logger LOGGER = Logger.getLogger(Utils.class.getName());
-  private static final String MAPS_API_KEY = "";
+  private static String MAPS_API_KEY;
+
+  static {
+    MAPS_API_KEY = "";
+    try {
+      MAPS_API_KEY = SecretHandler.getApiKey();
+    } catch (IOException e) {
+      LOGGER.warning(e.getMessage());
+    }
+  }
+
   private static final GeoApiContext context =
       new GeoApiContext.Builder().apiKey(MAPS_API_KEY).build();
