+++ src/main/java/com/google/sps/servlets/UserServlet.java
@@ -26,6 +26,7 @@ import com.google.appengine.api.datastore.Query.SortDirection;
 import com.google.appengine.api.users.UserService;
 import com.google.appengine.api.users.UserServiceFactory;
 import com.google.gson.Gson;
+import com.google.sps.Utils;
 import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collections;
