+++ src/test/java/com/google/sps/UserServletTest.java
@@ -32,7 +32,6 @@ import com.google.gson.Gson;
 import com.google.gson.reflect.TypeToken;
 import com.google.sps.servlets.EventServlet;
 import com.google.sps.servlets.UserServlet;
-import com.google.sps.servlets.Utils;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.io.StringWriter;
